CP-22600: Replace write_libxc_record by suspend_emu_manager
This changes the protocol that is followed between xenopsd and emu-manager, for
suspend/resume and migration, away from how xenguest does it.

The main difference is that, now, emu-manager has the initiative and decides on
the order on which things are done. Currently, emu-manager is given the
responsibility for the memory (xenguest) image, vGPU (DEMU) state, and when to
suspend the domain. It is likely that QEMU will be added to this list in the
near future. Xenopsd does not write and frame headers for these unless asked by
emu-manager.

Xenopsd starts emu-manager, specifying the main suspend fd, and optionally a
separate fd for the vGPU data, and then waits for a instructions from
emu-manager. A request for a new frame header comes in through a new
"prepare:x" message, where prepare is either "xenguest", or "vgpu" (and "qemu"
in future). Other messages are as before: "suspend", "result:x", "error:", and
the logging messages. Prepare and suspend are allowed in any order; result and
error always close the interaction with emu-manager, as well as the process
itself.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>